+++ app/src/main/java/uk/co/group35/app/user/DBModels/User.java
@@ -4,20 +4,20 @@ import org.springframework.data.annotation.Id;
 
 public class User {
     @Id
-    private Long id;
+    private Integer id;
     private String username;
     private String password;
 
-    public User(Long id, String username, String password) {
+    public User(Integer id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
